requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Document Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Image Processing
Pillow>=10.1.0
imagehash>=4.3.0

# Vector Database
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.36.0
torch>=2.1.0

# Vision Models (for captioning)
# Choose one based on GPU memory:
# - BLIP-2: transformers (already included)
# - Florence-2: transformers (already included)
# - Qwen2-VL: transformers (already included)

# Image Embeddings
open-clip-torch>=2.23.0

# Logging & Metrics
structlog>=23.2.0
python-json-logger>=2.0.0

# API Utilities
httpx>=0.25.0
pyngrok>=7.0.0
slowapi>=0.1.9

# Data Export
pyarrow>=14.0.0
pandas>=2.1.0

# Hybrid Search
rank-bm25>=0.2.2
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.11.0
ruff>=0.1.6
